requirements.txt
accelerate==0.28.0
faiss-cpu==1.7.4
hydra-core==1.3.2
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
sentencepiece==0.1.99
tensorboard==2.16.2
transformers==4.39.3
typer==0.12.3
wandb==0.16.4
fastapi==0.111.0
pydantic==2.7.4
